requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-cloud-discoveryengine>=0.11.0
google-cloud-storage
google-cloud-bigquery
python-dotenv
python-multipart
httpx
pypdf
python-docx
google-auth
google-auth-oauthlib
google-api-python-client
python-jose[cryptography]
cachetools

# Image Repository Pipeline dependencies
google-generativeai>=0.8.0
google-cloud-firestore>=2.19.0
google-cloud-secret-manager>=2.20.0
pillow>=10.4.0
pyyaml>=6.0.2

# Email Repository Pipeline dependencies
playwright>=1.41.0
google-genai>=1.0.0
